+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Support for custom rules plugins - all server plugins whose keys contain the phrase "delphi" are now retrieved and used for analysis
+
 ## [0.4.1] - 2023-08-08
 
 ### Changed